mv pydantic to config
